requirements.txt
torch==2.4.1
torchvision==0.19.1
pandas==2.0.3
Pillow==10.4.0
opencv-python-headless==4.10.0.84
PyYAML==6.0.2
matplotlib==3.7.5
